move api utils in separate file
